improves organization; adds node-pre-gyp
