+++ node/package.json
@@ -6,13 +6,15 @@
   "author": "Tiago Nascimento <tiago.nascimento@spruceid.com>",
   "license": "MIT",
   "dependencies": {
-    "neon-cli": "^0.5.3"
+    "neon-cli": "^0.5.3",
+    "node-pre-gyp": "^0.17.0"
   },
   "devDependencies": {
     "eslint": "^7.14.0",
     "eslint-config-prettier": "^7.0.0",
     "eslint-plugin-prettier": "^3.2.0",
     "jest": "^26.6.3",
+    "node-pre-gyp-github": "^1.4.3",
     "prettier": "2.2.1"
   },
   "scripts": {
@@ -32,6 +34,9 @@
   },
   "files": [
     "native/index.node",
+    "native/Cargo.toml",
+    "native/build.rs",
+    "native/src",
     "lib"
   ],
   "binary": {
